Moving DbDataReader creator to its own class
